+++ CHANGES.md
@@ -15,16 +15,22 @@ Changes in CUPS v2.4.3 (TBA)
 - Fixed a potential SNMP OID value overflow issue (Issue #431)
 - Fixed an OpenSSL certificate loading issue (Issue #465)
 - Fixed Brazilian Portuguese translations (Issue #288)
-- Fixed invalid memory access during generating IPP Everywhere queue (Issue #466)
+- Fixed invalid memory access during generating IPP Everywhere queue
+  (Issue #466)
 - Fixed memory leaks in `create_local_bg_thread()` (Issue #466)
+- Fixed media size tolerance in `ippeveprinter` (Issue #487)
+- Fixed `cupsd` default keychain location when building with OpenSSL
+  (Issue #529)
 - Fixed TLS certificate generation bugs.
 - `ippDeleteValues` would not delete the last value (Issue #556)
-- Ignore some of IPP defaults if the application sends its PPD alternative (Issue #484)
+- Ignore some of IPP defaults if the application sends its PPD alternative
+  (Issue #484)
 - Now look for default printer on network if needed (Issue #452)
 - Now report fax attributes and values as needed (Issue #459)
 - Now localize HTTP responses using the Content-Language value (Issue #426)
 - Raised file size limit for importing PPD via Web UI (Issue #433)
-- Update print-color-mode if the printer is modified via ColorModel PPD option (Issue #451)
+- Update print-color-mode if the printer is modified via ColorModel PPD option
+  (Issue #451)
 - Use localhost when printing via printer application (Issue #353)
 - Write defaults into /etc/cups/lpoptions if we're root (Issue #456)